Create README.md
This coursework project entails the study, design, implementation and testing of a responsive web application – a scholarship application portal for first degree holders from some selected states in Nigeria. The sample states of Nigeria used for this project include Abia, Anambra, Ebonyi, Enugu, and Imo states.
For the purpose of this documentation, the proposed name for the web application is “JacompPlus Online Scholarship Application Portal”; JOSAP for short.
For this coursework, the product owner of this project is proposed to be Jacomp+ – a tech academy that sponsors some Nigerian first degree holders to study computing technology courses abroad so as to promote the advancement of computing technology in Nigeria. Currently, the application process is done manually. This project digitizes the process by building a responsive web application that handles the registration/application process.
The proposed users of this web application system include the following:
•	The Visitor: 
This includes prospective scholarship applicants or anyone who visits the scholarship web portal to find some or more information about the scholarship programme. They have access to the “Home”, “Approved Universities And Courses”, and “Contact Us” pages of the website. They can also click on the “CREATE AN ACCOUNT” button to register if they wish to. They do not have access to other functionalities of the website, until they are registered.
•	The Registered Scholarship Applicants: 
These are registered applicants, and can log into the system with their respective login credentials to complete the scholarship application process. Some of the functionalities that they have access to include but not limited to include viewing their dashboards, creating their profiles to complete the scholarship application process, viewing their profiles, editing their profiles, uploading their credentials, logging out of the system, deleting their profiles, etc.
•	The Web Portal Administrator: 
This is the scholarship web portal administrator who has control over the activities going on in the system. He manages such functionalities such as adding new users (both admins and applicants), editing login credentials of users, viewing user profiles, updating user profiles, deleting user accounts, etc. 
For the purpose of this coursework, the following basic functionality should be considered and implemented:
•	The Visitor should be able to view the Home, Approved School and Contact Us pages, and also create an account as an applicant.
•	The Scholarship Applicants should be able to carry out all the activities performed by The Visitor (except creating another account with the same details used by him/her to create an account) and also view his/her dashboard (to complete the application process) and application details upon successful log in. He / She can also log out.
•	The Administrator should be able to have full control of the system; add/view/edit/update/delete users; read/delete messages from users, etc.

2.	Development Design, Architecture and Existing Code Libraries Used
This web application was developed on a Windows 10 x64-bit device using XAMPP as the development platform.
For the front-end development, HTML5 was used for the structure; CSS was used for styling, and JavaScript was used for some functionalities e.g. (Login/Register/Contact Us) Form validations, etc.
For the back-end development, PHP – for scripting, MySQL (phpMyAdmin) – for data storage, and Apache Web Server – for locally hosting the web application, were used.
The following external Bootstrap CSS (for styling) and JQuery/JavaScript (for functionalities) libraries were referenced and used for designing and developing this application:
•	https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css  
•	https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js
•	https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.0.0/css/bootstrap.min.css
•	https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js